[handlers]: add tests

+++ internal/api/handlers/models/movie.go
@@ -8,5 +8,5 @@ type Movie struct {
 	Title       string    `json:"title"`
 	Description string    `json:"description,omitempty"`
 	Date        time.Time `json:"date,omitempty"`
-	Rating      int       `json:"rating,omitempty"`
+	Rating      float32       `json:"rating,omitempty"`
 }
